requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
google-generativeai==0.7.2
python-dotenv==1.0.1
tzdata==2024.1
jsonschema==4.23.0
requests==2.32.3
